add ecomerce frontent and code new 2402

+++ ecommerce-app-clone/Models/Users.cs
@@ -9,7 +9,7 @@ namespace ecommerce_app_clone.Models
         public string LastName { get; set; }
         public string Password { get; set; }
         public string Email { get; set; }
-        public double Fund { get; set; }
+        public decimal Fund { get; set; }
         public string Type { get; set; }
         public int Status { get; set; }
         public DateTime CreateOn { get; set; }
